+++ src/main/java/me/drex/vanish/command/VanishCommand.java
@@ -1,16 +1,18 @@
 package me.drex.vanish.command;
 
+import com.mojang.authlib.GameProfile;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import me.drex.vanish.VanishMod;
 import me.drex.vanish.api.VanishAPI;
 import me.drex.vanish.config.ConfigManager;
+import me.drex.vanish.util.VanishManager;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.commands.CommandBuildContext;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
-import net.minecraft.commands.arguments.EntityArgument;
+import net.minecraft.commands.arguments.GameProfileArgument;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 
@@ -30,19 +32,19 @@ public class VanishCommand {
                         .executes(VanishCommand::reload)
                 ).then(
                     Commands.literal("on")
-                        .executes(ctx -> vanish(ctx.getSource(), true, Collections.singleton(ctx.getSource().getPlayerOrException())))
+                        .executes(ctx -> vanish(ctx.getSource(), true, Collections.singleton(ctx.getSource().getPlayerOrException().getGameProfile())))
                         .then(
-                            Commands.argument("players", EntityArgument.players())
+                            Commands.argument("players", GameProfileArgument.gameProfile())
                                 .requires(src -> Permissions.check(src, "vanish.command.vanish.other", 2))
-                                .executes(ctx -> vanish(ctx.getSource(), true, EntityArgument.getPlayers(ctx, "players")))
+                                .executes(ctx -> vanish(ctx.getSource(), true, GameProfileArgument.getGameProfiles(ctx, "players")))
                         )
                 ).then(
                     Commands.literal("off")
-                        .executes(ctx -> vanish(ctx.getSource(), false, Collections.singleton(ctx.getSource().getPlayerOrException())))
+                        .executes(ctx -> vanish(ctx.getSource(), false, Collections.singleton(ctx.getSource().getPlayerOrException().getGameProfile())))
                         .then(
-                            Commands.argument("players", EntityArgument.players())
+                            Commands.argument("players", GameProfileArgument.gameProfile())
                                 .requires(src -> Permissions.check(src, "vanish.command.vanish.other", 2))
-                                .executes(ctx -> vanish(ctx.getSource(), false, EntityArgument.getPlayers(ctx, "players")))
+                                .executes(ctx -> vanish(ctx.getSource(), false, GameProfileArgument.getGameProfiles(ctx, "players")))
                         )
                 )
         );
@@ -51,7 +53,7 @@ public class VanishCommand {
     public static int vanish(CommandContext<CommandSourceStack> ctx) throws CommandSyntaxException {
         ServerPlayer player = ctx.getSource().getPlayerOrException();
         boolean vanished = VanishAPI.isVanished(player);
-        return vanish(ctx.getSource(), !vanished, Collections.singleton(player));
+        return vanish(ctx.getSource(), !vanished, Collections.singleton(player.getGameProfile()));
     }
 
     public static int reload(CommandContext<CommandSourceStack> ctx) {
@@ -66,15 +68,15 @@ public class VanishCommand {
         }
     }
 
-    public static int vanish(CommandSourceStack src, boolean vanish, Collection<ServerPlayer> targets) throws CommandSyntaxException {
+    public static int vanish(CommandSourceStack src, boolean vanish, Collection<GameProfile> targets) throws CommandSyntaxException {
         int result = 0;
-        for (ServerPlayer target : targets) {
-            if (!VanishAPI.setVanish(target, vanish)) continue;
-            if (src.getPlayerOrException() == target) {
+        for (GameProfile target : targets) {
+            if (!VanishManager.setVanished(target, src.getServer(), vanish)) continue;
+            ServerPlayer player = src.getPlayer();
+            if (player != null && player.getGameProfile().equals(target)) {
                 src.sendSuccess(() -> Component.translatable(vanish ? "text.vanish.command.vanish.enable" : "text.vanish.command.vanish.disable"), false);
             } else {
-                src.sendSuccess(() -> Component.translatable(vanish ? "text.vanish.command.vanish.enable.other" : "text.vanish.command.vanish.disable.other", target.getDisplayName()), false);
-                target.sendSystemMessage(Component.translatable(vanish ? "text.vanish.command.vanish.enable" : "text.vanish.command.vanish.disable"));
+                src.sendSuccess(() -> Component.translatable(vanish ? "text.vanish.command.vanish.enable.other" : "text.vanish.command.vanish.disable.other", target.getName()), false);
             }
             result++;
         }